Split Files.js into two modules
