+++ client/src/js/files/components/Manager.js
@@ -1,62 +1,14 @@
 import React from "react";
-import PropTypes from "prop-types";
 import Dropzone from "react-dropzone";
 import { push } from "react-router-redux";
 import { capitalize, filter } from "lodash";
 import { connect } from "react-redux";
-import { Col, ListGroup, Row } from "react-bootstrap";
+import { ListGroup } from "react-bootstrap";
 
-import { byteSize, createRandomString } from "../../utils";
+import File from "./File";
+import { createRandomString } from "../../utils";
 import { findFiles, removeFile, upload, uploadProgress } from "../actions";
-import { Button, Icon, ListGroupItem, LoadingPlaceholder, Pagination, RelativeTime, ViewHeader } from "../../base";
-
-const File = (props) => {
-    let creation;
-
-    if (props.user === null) {
-        creation = (
-            <span>
-                Retrieved <RelativeTime time={props.uploaded_at} />
-            </span>
-        );
-    } else {
-        creation = <span>Uploaded <RelativeTime time={props.uploaded_at} /> by {props.user.id}</span>;
-    }
-
-    return (
-        <ListGroupItem className="spaced">
-            <Row>
-                <Col md={5}>
-                    <strong>{props.name}</strong>
-                </Col>
-                <Col md={2}>
-                    {byteSize(props.size)}
-                </Col>
-                <Col md={4}>
-                    {creation}
-                </Col>
-                <Col md={1}>
-                    <Icon
-                        name="remove"
-                        bsStyle="danger"
-                        style={{fontSize: "17px"}}
-                        pullRight onClick={() => props.onRemove(props.id)}
-                    />
-                </Col>
-            </Row>
-        </ListGroupItem>
-    );
-};
-
-File.propTypes = {
-    id: PropTypes.string,
-    name: PropTypes.string,
-    size: PropTypes.number,
-    file: PropTypes.object,
-    uploaded_at: PropTypes.string,
-    user: PropTypes.object,
-    onRemove: PropTypes.func
-};
+import { Button, Icon, ListGroupItem, LoadingPlaceholder, Pagination, ViewHeader } from "../../base";
 
 class FileManager extends React.Component {
 
@@ -134,7 +86,16 @@ class FileManager extends React.Component {
     }
 }
 
-const mapStateToProps = (state) => ({...state.files});
+const mapStateToProps = (state) => {
+    const { documents, page, found_count, total_count } = state.files;
+
+    return {
+        documents,
+        page,
+        found_count,
+        total_count
+    };
+};
 
 const mapDispatchProps = (dispatch) => ({
