+++ client/src/js/subtraction/components/Subtraction.js
@@ -3,13 +3,13 @@ import { Switch, Route } from "react-router-dom";
 
 import SubtractionList from "./List";
 import SubtractionDetail from "./Detail";
-import Files from "../../files/components/Files";
+import FileManager from "../../files/components/Manager";
 
 const Subtraction = () => (
     <div className="container">
         <Switch>
             <Route path="/subtraction" component={SubtractionList} exact />
-            <Route path="/subtraction/files" render={() => <Files fileType="subtraction" />} />
+            <Route path="/subtraction/files" render={() => <FileManager fileType="subtraction" />} />
             <Route path="/subtraction/:subtractionId" component={SubtractionDetail} />
         </Switch>
     </div>
